Dop 1643 (#360)
* Paginated S3 Reading

Also cleans up orphan .processed files, and limits number of object
processing routines

+++ main.go
@@ -5,7 +5,6 @@ import (
 	"context"
 	"slices"
 	"strings"
-	"sync"
 	"time"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
@@ -18,6 +17,7 @@ import (
 	"github.com/quortex/influxdb-athena-crawler/pkg/influxdb"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
+	"golang.org/x/sync/errgroup"
 )
 
 var opts flags.Options
@@ -55,18 +55,26 @@ func main() {
 	}
 	s3Cli := s3.NewFromConfig(cfg)
 
-	// List objects matching bucket / prefix
-	res, err := s3Cli.ListObjects(ctx, &s3.ListObjectsInput{
+	var elems s3.ListObjectsOutput
+
+	p := s3.NewListObjectsV2Paginator(s3Cli, &s3.ListObjectsV2Input{
 		Bucket: aws.String(opts.Bucket),
 		Prefix: aws.String(opts.Prefix),
 	})
-	if err != nil {
-		log.Fatal().Err(err).Msg("Unable to list objects")
+
+	for p.HasMorePages() {
+		page, err := p.NextPage(ctx)
+		if err != nil {
+			log.Fatal().Err(err).Msg("Unable to get object page")
+		}
+
+		// Log the objects found
+		elems.Contents = slices.Concat(elems.Contents, page.Contents)
 	}
 
-	unprocCsvs, procCsvs := filterBucketContent(*res, opts.Suffix, opts.ProcessedFlagSuffix)
+	unprocCsvs, procCsvs, orphanFlags := filterBucketContent(elems, opts.Suffix, opts.ProcessedFlagSuffix)
 
-	if len(procCsvs.Contents)+len(unprocCsvs.Contents) == 0 {
+	if len(procCsvs.Contents)+len(unprocCsvs.Contents)+len(orphanFlags.Contents) == 0 {
 		log.Info().Msg("No objects matching bucket / prefix, processing done !")
 		return
 	}
@@ -87,19 +95,30 @@ func main() {
 	defer influxWriter.Close()
 
 	if len(unprocCsvs.Contents) > 0 {
-		parallelApply(unprocCsvs, func(o types.Object) {
-			if err := processObject(ctx, dwn, upl, influxWriter, o); err != nil {
-				log.Error().Err(err).Msg("Processing error")
-			}
+		err = parallelApply(ctx, unprocCsvs, func(o types.Object) error {
+			return processObject(ctx, dwn, upl, influxWriter, o)
 		})
+		if err != nil {
+			log.Fatal().Err(err).Msg("Failed processing objects")
+		}
 	}
 
 	if opts.CleanObjects && len(procCsvs.Contents) > 0 {
-		parallelApply(procCsvs, func(o types.Object) {
-			if err := cleanObject(ctx, s3Cli, o); err != nil {
-				log.Error().Err(err).Msg("Cleaning error")
-			}
+		err = parallelApply(ctx, procCsvs, func(o types.Object) error {
+			return cleanObject(ctx, s3Cli, o)
 		})
+		if err != nil {
+			log.Fatal().Err(err).Msg("Failed cleaning objects")
+		}
+	}
+
+	if len(orphanFlags.Contents) > 0 {
+		err = parallelApply(ctx, orphanFlags, func(o types.Object) error {
+			return cleanObject(ctx, s3Cli, o)
+		})
+		if err != nil {
+			log.Fatal().Err(err).Msg("Failed cleaning orphan flags")
+		}
 	}
 
 	log.Info().
@@ -107,46 +126,61 @@ func main() {
 		Msg("Processing ended !")
 }
 
-func filterBucketContent(elems s3.ListObjectsOutput, csvSuffix, processedFlagSuffix string) (unprocessed, processed s3.ListObjectsOutput) {
+func filterBucketContent(elems s3.ListObjectsOutput, csvSuffix, processedFlagSuffix string) (unprocessed, processed, orphanFlags s3.ListObjectsOutput) {
 	// Rely on .processed files present on the bucket to detect which csv
 	// has already been pushed to influx and which has yet to be processed
+	// List .processed files that do not match any data file in order to clean them up, this can happen if the crawler was interrupted
 	csvFiles := []types.Object{}
-	processedElems := []string{}
+	flags := []types.Object{}
+	objectNames := []string{}
+	flagNames := []string{}
 
 	if len(csvSuffix) == 0 {
-		return unprocessed, processed
+		return unprocessed, processed, orphanFlags
 	}
 	for _, s := range elems.Contents {
 		if strings.HasSuffix(*s.Key, csvSuffix) {
 			csvFiles = append(csvFiles, s)
+			objectNames = append(objectNames, *s.Key)
 		}
 		if strings.HasSuffix(*s.Key, processedFlagSuffix) {
-			processedElems = append(processedElems, strings.ReplaceAll(*s.Key, processedFlagSuffix, csvSuffix))
+			flags = append(flags, s)
+			flagNames = append(flagNames, *s.Key)
 		}
 	}
 
 	for _, o := range csvFiles {
-		if !slices.Contains(processedElems, *o.Key) {
+		flagName := strings.ReplaceAll(*o.Key, csvSuffix, processedFlagSuffix)
+		if !slices.Contains(flagNames, flagName) {
 			unprocessed.Contents = append(unprocessed.Contents, o)
 		} else {
 			processed.Contents = append(processed.Contents, o)
 		}
 	}
 
-	return unprocessed, processed
+	for _, f := range flags {
+		fileName := strings.ReplaceAll(*f.Key, processedFlagSuffix, csvSuffix)
+		if !slices.Contains(objectNames, fileName) {
+			orphanFlags.Contents = append(orphanFlags.Contents, f)
+		}
+
+	}
+
+	return unprocessed, processed, orphanFlags
 }
 
-func parallelApply(list s3.ListObjectsOutput, fn func(o types.Object)) {
-	var wg sync.WaitGroup
-	wg.Add(len(list.Contents))
+func parallelApply(ctx context.Context, list s3.ListObjectsOutput, fn func(o types.Object) error) error {
+	//Limit the number of parallel routines doing the processing.
+	g, _ := errgroup.WithContext(ctx)
+	g.SetLimit(opts.MaxRoutines)
+
 	for _, item := range list.Contents {
 		o := item
-		go func() {
-			defer wg.Done()
-			fn(o)
-		}()
+		g.Go(func() error {
+			return fn(o)
+		})
 	}
-	wg.Wait()
+	return g.Wait()
 }
 
 func processObject(
@@ -238,18 +272,20 @@ func cleanObject(
 			return err
 		}
 
-		markerFileName := strings.ReplaceAll(*o.Key, opts.Suffix, opts.ProcessedFlagSuffix)
-		_, err = s3Cli.DeleteObject(ctx, &s3.DeleteObjectInput{
-			Bucket: aws.String(opts.Bucket),
-			Key:    &markerFileName,
-		})
-		if err != nil {
-			log.Error().
-				Err(err).
-				Str("bucket", opts.Bucket).
-				Str("object", aws.ToString(o.Key)).
-				Msg("Unable to delete object")
-			return err
+		if !strings.Contains(*o.Key, opts.ProcessedFlagSuffix) {
+			markerFileName := strings.ReplaceAll(*o.Key, opts.Suffix, opts.ProcessedFlagSuffix)
+			_, err = s3Cli.DeleteObject(ctx, &s3.DeleteObjectInput{
+				Bucket: aws.String(opts.Bucket),
+				Key:    &markerFileName,
+			})
+			if err != nil {
+				log.Error().
+					Err(err).
+					Str("bucket", opts.Bucket).
+					Str("object", aws.ToString(o.Key)).
+					Msg("Unable to delete object")
+				return err
+			}
 		}
 	}
 	return nil
